refactor: reorganize similar types of files
