+++ packages/effects/source/combinators/index.ts
@@ -1,0 +1,2 @@
+export * from './combine'
+export * from './race'
