+++ packages/effects/source/factories/get.ts
@@ -1,5 +1,5 @@
 import { withEnv } from '@typed/env'
 import { id } from '@typed/lambda'
-import { Effect } from './Effect'
+import { Effect } from '../Effect'
 
 export const get = <A>(): Effect<A, A> => Effect.fromEnv<A, A>(withEnv(id))
