+++ packages/effects/source/failures/Failure.ts
@@ -1,11 +1,12 @@
 import { Either } from '@typed/either'
 import { Resume } from '@typed/env'
-import { Maybe, unpack } from '@typed/maybe'
+import { Maybe, Nothing, unpack } from '@typed/maybe'
 
 export type Fail<Err> = { readonly failure: <A>(error: Err) => Resume<Failure<Err, A>> }
 
 export class Failure<Err, A> {
-  public static of = <Err, A>(error: Err, value: Maybe<A>) => new Failure<Err, A>(error, value)
+  public static of = <Err, A>(error: Err, value: Maybe<A> = Nothing) =>
+    new Failure<Err, A>(error, value)
 
   constructor(readonly error: Err, readonly value: Maybe<A>) {}
