+++ packages/effects/source/failures/fail.ts
@@ -1,8 +1,8 @@
 import { CapabilitiesOf, Provide, Resume } from '@typed/env'
 import { Just } from '@typed/maybe'
-import { Effect, Return, Yield } from './Effect'
+import { Effect, Return, Yield } from '../Effect'
+import { runWith } from '../run'
 import { Fail, Failure } from './Failure'
-import { runWith } from './runWith'
 
 export function* fail<Err>(error: Err): Effect<Fail<Err>, void> {
   yield ({ failure }) => failure(error)
