+++ packages/effects/source/index.ts
@@ -1,10 +1,6 @@
-export * from './combine'
+export * from './combinators'
+export * from './factories'
+export * from './failures'
+export * from './fibers'
+export * from './run'
 export * from './Effect'
-export * from './fail'
-export * from './Failure'
-export * from './get'
-export * from './race'
-export * from './runEffect'
-export * from './runEffects'
-export * from './runWith'
-export * from './startEffect'
