+++ packages/effects/source/run/startEffect.ts
@@ -1,5 +1,5 @@
 import { Env, Resume } from '@typed/env'
-import { Capabilities, Effect } from './Effect'
+import { Capabilities, Effect } from '../Effect'
 
 export const startEffect = <A extends Effect<any, any>>(effect: A): Env<Capabilities<A>, A> => (
   _: Capabilities<A>,
